refactor: mejorar semántica de errores REST y renombrar middleware de métodos no permitidos
