+++ tp-final-integrador/src/middlewares/error.middleware.js
@@ -30,14 +30,24 @@ export const globalErrorHandler = (err, req, res, _next) => {
     });
   }
 
-  // 2. Si llegamos acá, es un BUG (Programming Error) o error no controlado
+  // 2. Errores de sintaxis de JSON (específicos de express.json)
+  if (err instanceof SyntaxError && err.status === 400 && 'body' in err) {
+    return errorResponse({
+      res,
+      errorType: ERROR_CODES.BAD_REQUEST,
+      message: 'El cuerpo de la petición (JSON) tiene un formato inválido',
+      details: process.env.NODE_ENV === 'development' ? [{ error: err.message }] : [],
+    });
+  }
+
+  // 3. BUG (Programming Error) o error no controlado
   console.error('ERROR NO CONTROLADO:', err);
 
   const status = err.status || 500;
   const message = status === 500 ? 'Error interno del servidor' : err.message;
 
   // En desarrollo mostramos el stack trace para debuguear mejor
-  const details = process.env.NODE_ENV === 'development' ? { stack: err.stack } : [];
+  const details = process.env.NODE_ENV === 'development' ? [{ stack: err.stack }] : [];
 
   return errorResponse({
     res,
